works properly now
